+++ tests/test_algorithms.py
@@ -1,158 +1,185 @@
+from catanatron.game import Game
 from catanatron.algorithms import longest_road
-from catanatron.models.board import Board
-from catanatron.models.board_initializer import EdgeRef, NodeRef
-from catanatron.models.player import Color
-
-# ===== Buildable nodes
-def test_buildable_nodes():
-    board = Board()
-    nodes = board.buildable_nodes(Color.RED)
-    assert len(nodes) == 0
-    nodes = board.buildable_nodes(Color.RED, initial_build_phase=True)
-    assert len(nodes) == 54
+from catanatron.models.board_initializer import NodeRef, EdgeRef
+from catanatron.models.actions import Action, ActionType
+from catanatron.models.player import SimplePlayer, Color
+from catanatron.models.decks import ResourceDeck
 
 
-def test_placing_settlement_removes_four_buildable_nodes():
-    board = Board()
-    board.build_settlement(
-        Color.RED, board.nodes[((0, 0, 0), NodeRef.SOUTH)], initial_build_phase=True
+def road(player, node_id):
+    """Helper function to create BUILD_ROAD actions"""
+    return Action(player, ActionType.BUILD_ROAD, node_id)
+
+
+def test_longest_road_simple():
+    red = SimplePlayer(Color.RED)
+    blue = SimplePlayer(Color.BLUE)
+    red.resource_deck += ResourceDeck.starting_bank()
+    blue.resource_deck += ResourceDeck.starting_bank()
+
+    game = Game(players=[red, blue])
+    nodes = game.board.nodes
+    edges = game.board.edges
+
+    game.execute(
+        Action(
+            red, ActionType.BUILD_FIRST_SETTLEMENT, nodes[((0, 0, 0), NodeRef.SOUTH)].id
+        )
     )
-    nodes = board.buildable_nodes(Color.RED)
-    assert len(nodes) == 0
-    nodes = board.buildable_nodes(Color.RED, initial_build_phase=True)
-    assert len(nodes) == 50
-    nodes = board.buildable_nodes(Color.BLUE, initial_build_phase=True)
-    assert len(nodes) == 50
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.SOUTHEAST)].id))
+
+    color, path = longest_road(game.board, game.players, game.actions)
+    assert color is None
+
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.EAST)].id))
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.NORTHEAST)].id))
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.NORTHWEST)].id))
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.WEST)].id))
+
+    color, path = longest_road(game.board, game.players, game.actions)
+    assert color == Color.RED
+    assert len(path) == 5
 
 
-def test_buildable_nodes_respects_distance_two():
-    board = Board()
-    board.build_settlement(
-        Color.RED, board.nodes[((0, 0, 0), NodeRef.SOUTH)], initial_build_phase=True
+def test_longest_road_tie():
+    red = SimplePlayer(Color.RED)
+    blue = SimplePlayer(Color.BLUE)
+    red.resource_deck += ResourceDeck.starting_bank()
+    blue.resource_deck += ResourceDeck.starting_bank()
+
+    game = Game(players=[red, blue])
+    nodes = game.board.nodes
+    edges = game.board.edges
+
+    game.execute(
+        Action(
+            red,
+            ActionType.BUILD_FIRST_SETTLEMENT,
+            nodes[((0, 0, 0), NodeRef.SOUTH)].id,
+        ),
     )
-
-    board.build_road(Color.RED, board.edges[((0, 0, 0), EdgeRef.SOUTHWEST)])
-    nodes = board.buildable_nodes(Color.RED)
-    assert len(nodes) == 0
-
-    board.build_road(Color.RED, board.edges[((0, 0, 0), EdgeRef.WEST)])
-    nodes = board.buildable_nodes(Color.RED)
-    assert len(nodes) == 1
-    assert nodes.pop() == board.nodes[((0, 0, 0), NodeRef.NORTHWEST)]
-
-
-def test_cant_use_enemy_roads_to_connect():
-    board = Board()
-    board.build_settlement(
-        Color.RED, board.nodes[((0, 0, 0), NodeRef.SOUTH)], initial_build_phase=True
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.SOUTHEAST)].id))
+    game.execute(
+        Action(
+            blue,
+            ActionType.BUILD_FIRST_SETTLEMENT,
+            nodes[((0, 2, -2), NodeRef.SOUTH)].id,
+        ),
     )
-    board.build_road(Color.RED, board.edges[((0, 0, 0), EdgeRef.SOUTHEAST)])
+    game.execute(road(blue, edges[((0, 2, -2), EdgeRef.SOUTHEAST)].id))
 
-    board.build_settlement(
-        Color.BLUE,
-        board.nodes[((0, 0, 0), NodeRef.NORTHEAST)],
-        initial_build_phase=True,
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.EAST)].id))
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.NORTHEAST)].id))
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.NORTHWEST)].id))
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.WEST)].id))
+
+    game.execute(road(blue, edges[((0, 2, -2), EdgeRef.EAST)].id))
+    game.execute(road(blue, edges[((0, 2, -2), EdgeRef.NORTHEAST)].id))
+    game.execute(road(blue, edges[((0, 2, -2), EdgeRef.NORTHWEST)].id))
+    game.execute(road(blue, edges[((0, 2, -2), EdgeRef.WEST)].id))
+
+    color, path = longest_road(game.board, game.players, game.actions)
+    assert color == Color.RED  # even if blue also has 5-road. red had it first
+    assert len(path) == 5
+
+    game.execute(road(blue, edges[((0, 2, -2), EdgeRef.SOUTHWEST)].id))
+    color, path = longest_road(game.board, game.players, game.actions)
+    assert color == Color.BLUE
+    assert len(path) == 6
+
+
+# test: complicated circle around
+def test_complicated_road():  # classic 8-like roads
+    red = SimplePlayer(Color.RED)
+    blue = SimplePlayer(Color.BLUE)
+    red.resource_deck += ResourceDeck.starting_bank()
+    blue.resource_deck += ResourceDeck.starting_bank()
+
+    game = Game(players=[red, blue])
+    nodes = game.board.nodes
+    edges = game.board.edges
+
+    game.execute(
+        Action(
+            red, ActionType.BUILD_FIRST_SETTLEMENT, nodes[((0, 0, 0), NodeRef.SOUTH)].id
+        ),
     )
-    board.build_road(Color.BLUE, board.edges[((0, 0, 0), EdgeRef.EAST)])
-    board.build_road(Color.BLUE, board.edges[((0, 0, 0), EdgeRef.NORTHEAST)])
-    board.build_road(Color.BLUE, board.edges[((1, 0, -1), EdgeRef.WEST)])
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.SOUTHEAST)].id))
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.EAST)].id))
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.NORTHEAST)].id))
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.NORTHWEST)].id))
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.WEST)].id))
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.SOUTHWEST)].id))
 
-    nodes = board.buildable_nodes(Color.RED)
-    assert len(nodes) == 0
+    game.execute(road(red, edges[((1, -1, 0), EdgeRef.SOUTHWEST)].id))
+    game.execute(road(red, edges[((1, -1, 0), EdgeRef.SOUTHEAST)].id))
+    game.execute(road(red, edges[((1, -1, 0), EdgeRef.EAST)].id))
+    game.execute(road(red, edges[((1, -1, 0), EdgeRef.NORTHEAST)].id))
+    game.execute(road(red, edges[((1, -1, 0), EdgeRef.NORTHWEST)].id))
 
-    nodes = board.buildable_nodes(Color.BLUE)
-    assert len(nodes) == 1
+    color, path = longest_road(game.board, game.players, game.actions)
+    assert color == Color.RED
+    assert len(path) == 11
+
+    game.execute(road(red, edges[((2, -2, 0), EdgeRef.SOUTHWEST)].id))
+
+    color, path = longest_road(game.board, game.players, game.actions)
+    assert color == Color.RED
+    assert len(path) == 11
 
 
-# ===== Buildable edges
-def test_buildable_edges_simple():
-    board = Board()
-    board.build_settlement(
-        Color.RED, board.nodes[((0, 0, 0), NodeRef.SOUTH)], initial_build_phase=True
+def test_triple_longest_road_tie():
+    red = SimplePlayer(Color.RED)
+    blue = SimplePlayer(Color.BLUE)
+    white = SimplePlayer(Color.WHITE)
+    red.resource_deck += ResourceDeck.starting_bank()
+    blue.resource_deck += ResourceDeck.starting_bank()
+    white.resource_deck += ResourceDeck.starting_bank()
+
+    game = Game(players=[red, blue, white])
+    nodes = game.board.nodes
+    edges = game.board.edges
+
+    game.execute(
+        Action(
+            red, ActionType.BUILD_FIRST_SETTLEMENT, nodes[((0, 0, 0), NodeRef.SOUTH)].id
+        ),
     )
-    buildable = board.buildable_edges(Color.RED)
-    assert len(buildable) == 3
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.SOUTHEAST)].id))
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.EAST)].id))
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.NORTHEAST)].id))
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.NORTHWEST)].id))
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.WEST)].id))
+    game.execute(road(red, edges[((0, 0, 0), EdgeRef.SOUTHWEST)].id))
 
-
-def test_buildable_edges():
-    board = Board()
-    board.build_settlement(
-        Color.RED, board.nodes[((0, 0, 0), NodeRef.SOUTH)], initial_build_phase=True
+    game.execute(
+        Action(
+            blue,
+            ActionType.BUILD_FIRST_SETTLEMENT,
+            nodes[((-2, 2, 0), NodeRef.SOUTH)].id,
+        ),
     )
-    board.build_road(Color.RED, board.edges[((0, 0, 0), EdgeRef.SOUTHWEST)])
-    buildable = board.buildable_edges(Color.RED)
-    assert len(buildable) == 4
+    game.execute(road(blue, edges[((-2, 2, 0), EdgeRef.SOUTHEAST)].id))
+    game.execute(road(blue, edges[((-2, 2, 0), EdgeRef.EAST)].id))
+    game.execute(road(blue, edges[((-2, 2, 0), EdgeRef.NORTHEAST)].id))
+    game.execute(road(blue, edges[((-2, 2, 0), EdgeRef.NORTHWEST)].id))
+    game.execute(road(blue, edges[((-2, 2, 0), EdgeRef.WEST)].id))
+    game.execute(road(blue, edges[((-2, 2, 0), EdgeRef.SOUTHWEST)].id))
 
-
-def test_water_edge_is_not_buildable():
-    board = Board()
-    board.build_settlement(
-        Color.RED, board.nodes[((0, 2, -2), NodeRef.NORTH)], initial_build_phase=True
+    game.execute(
+        Action(
+            white,
+            ActionType.BUILD_FIRST_SETTLEMENT,
+            nodes[((2, -2, 0), NodeRef.SOUTH)].id,
+        ),
     )
-    buildable = board.buildable_edges(Color.RED)
-    assert len(buildable) == 2
+    game.execute(road(white, edges[((2, -2, 0), EdgeRef.SOUTHEAST)].id))
+    game.execute(road(white, edges[((2, -2, 0), EdgeRef.EAST)].id))
+    game.execute(road(white, edges[((2, -2, 0), EdgeRef.NORTHEAST)].id))
+    game.execute(road(white, edges[((2, -2, 0), EdgeRef.NORTHWEST)].id))
+    game.execute(road(white, edges[((2, -2, 0), EdgeRef.WEST)].id))
+    game.execute(road(white, edges[((2, -2, 0), EdgeRef.SOUTHWEST)].id))
 
-
-# ===== Find connected components
-def test_connected_components_empty_board():
-    board = Board()
-    components = board.find_connected_components(Color.RED)
-    assert len(components) == 0
-
-
-def test_one_connected_component():
-    board = Board()
-    board.build_settlement(
-        Color.RED, board.nodes[((0, 0, 0), NodeRef.SOUTH)], initial_build_phase=True
-    )
-    board.build_road(Color.RED, board.edges[((0, 0, 0), EdgeRef.SOUTHEAST)])
-    board.build_settlement(
-        Color.RED, board.nodes[((0, 0, 0), NodeRef.NORTHEAST)], initial_build_phase=True
-    )
-    board.build_road(Color.RED, board.edges[((0, 0, 0), EdgeRef.EAST)])
-    components = board.find_connected_components(Color.RED)
-    assert len(components) == 1
-
-    board.build_road(Color.RED, board.edges[((0, 0, 0), EdgeRef.NORTHEAST)])
-    components = board.find_connected_components(Color.RED)
-    assert len(components) == 1
-
-
-def test_two_connected_components():
-    board = Board()
-    board.build_settlement(
-        Color.RED, board.nodes[((0, 0, 0), NodeRef.SOUTH)], initial_build_phase=True
-    )
-    board.build_road(Color.RED, board.edges[((0, 0, 0), EdgeRef.SOUTHWEST)])
-    components = board.find_connected_components(Color.RED)
-    assert len(components) == 1
-
-    board.build_settlement(
-        Color.RED, board.nodes[((0, 0, 0), NodeRef.NORTHEAST)], initial_build_phase=True
-    )
-    board.build_road(Color.RED, board.edges[((0, 0, 0), EdgeRef.NORTHEAST)])
-    components = board.find_connected_components(Color.RED)
-    assert len(components) == 2
-
-
-def test_three_connected_components_bc_enemy_cut_road():
-    board = Board()
-    board.build_settlement(
-        Color.RED, board.nodes[((0, 0, 0), NodeRef.SOUTH)], initial_build_phase=True
-    )
-    board.build_road(Color.RED, board.edges[((0, 0, 0), EdgeRef.SOUTHWEST)])
-
-    board.build_settlement(
-        Color.RED, board.nodes[((0, 0, 0), NodeRef.NORTHEAST)], initial_build_phase=True
-    )
-    board.build_road(Color.RED, board.edges[((0, 0, 0), EdgeRef.NORTHEAST)])
-    board.build_road(Color.RED, board.edges[((0, 0, 0), EdgeRef.NORTHWEST)])
-    board.build_road(Color.RED, board.edges[((-1, 1, 0), EdgeRef.NORTHEAST)])
-
-    board.build_settlement(
-        Color.BLUE,
-        board.nodes[((0, 0, 0), NodeRef.NORTHWEST)],
-        initial_build_phase=True,
-    )
-    components = board.find_connected_components(Color.RED)
-    assert len(components) == 3
+    color, path = longest_road(game.board, game.players, game.actions)
+    assert color == Color.RED
+    assert len(path) == 6
